Updated code to read a single 6 digit number from xlsx file
I created a function called digit_extraction.R that reads a single
column of code containing 6 digits and converts it to a format that our
code is able to read. I modified the load_4CExcelwb.R function to call
this function. I modified QuantWriting master script.R to
automatically source digit_extraction.R. I also modified it to load
"4CAssignmentMaster_extract_individual_digits.xlsx", which I was using
to test this new format of input. I included this xlsx file in the
commit.